Fix: 프래그먼트 연결 오류
