+++ app/src/main/res/layout/fragment_timetable.xml
@@ -20,7 +20,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:textStyle="bold"
-            android:textSize="35sp"
+            android:textSize="25sp"
             android:textColor="@color/black"
             android:text="@string/timetable_team"
             app:layout_constraintTop_toTopOf="parent"
@@ -36,7 +36,7 @@
             app:layout_constraintEnd_toStartOf="@id/btn_my_timetable"
             app:layout_constraintTop_toTopOf="parent"
             app:layout_constraintBottom_toBottomOf="parent"
-            android:layout_marginHorizontal="30dp"/>
+            android:layout_marginHorizontal="20dp"/>
 
         <ImageButton
             android:id="@+id/btn_my_timetable"
