+++ app/src/main/res/layout/item_rv_group_list.xml
@@ -17,7 +17,6 @@
             android:id="@+id/team_img"
             android:layout_width="45dp"
             android:layout_height="45dp"
-            android:src="@drawable/img_team_profile"
             app:civ_border_overlay="true"
             app:civ_border_width="1dp"
             app:civ_border_color="@color/black"
